Initial server-independent version
The aim is to be able to have 2 different client implementations which
perform the same actions. The first is a local one, with all existing
functionality. The second one is a remote one where the results are
transmitted from a cluster which the comparator cannot access.

This is useful in situations where there is no node which can see both
clusters. For example, if one runs on bare metal in a firewalled data
center and the other runs on the cloud with no public IP address access.