Cached remote proxy
- Added a new parameter `remoteCacheSize` to allow the client to cache a
number of objects. For example, if this parameters is set to 1000, the
client will ask the remote server for 1000 records instead of 1. This is
done in a new thread, so the client requesting 1 record will get the
first record returned as soon as it's returned, with the filling thread
then populating the rest into a queue, until either the queue is filled
or there are no more records on the server. When the client reads (or a
50ms timer elapses), the filling thread will try to read more records
from the server, keeping the cache filled.

- Added extra logging in when an IOException occurs.