Added hashing of records between remote servers.
Added a parameter which defaults to true to allow hash comparisons when
using a remote server for comparisons. This will compute the hash of the
record locally and simply send the hash to the comparator controller.
This is very efficient, especially in cases where the compare mode is
RECORDS_DIFFERENT or where there are small numbers of records likely to
be different when using RECORD_DIFFERENCES.