v 0.9.8Fixed order dependence in maps
When using a remote server and comparing records, an initial pass is
performed on the records by using a hash. This works great, except when
it comes to maps where the hash will be determined by the storage order,
even though the contents might be identical. This can lead to erroneous
messages in RECORDS_DIFFERENT mode when the hash is the only thing used.

Hence, in this mode with a remote server using hashes, the maps are by
default sorted into ordered lists using Aerospike's ordering before
hashing. This will cause additional CPU utilization.

Introduced the --sortMaps flag to override this behaviour if desired.