Added a "rerun" mode
Sometimes differences are noticed between clusters because Aerospike's
XDR is still shipping records. These are transient differences and
should not appear in the next run. Using "rerun" allows the comparator
to just run on the records which were found to be different on the
original run.

This mode requires both an input file (the output of the previous run)
and an output file (the results it's found this run). Hence, the `-f
filename` parameter has been removed and replaced with `-i inputFile`
and `-o outputFile`.

Additionally, the format of the output has been revised to make the
summary easier to read.