Added namespace mapping
For situations where there are the same sets and same data in different
namespaces, the ability to map namespaces across clusters has been
added. For example, staging and prod clusters might contain the same
data, but might have namespaces called "stagingShared" and "prodShared"
respectively. This mapping feature gives the comparator the ability to
handle this case.

Updated the README.md file to reflect this.

Also fixed a null pointer exception when the first cluster in a
comparison was missing the record but it was present in other clusters.