Fixed re-runs on clusters with different names
When multiple clusters have mappings from one namespace to a different
namespace, the different namespace can appear in the output file. For
example, NS A on cluster 1 is known as NS B on cluster 2, so A->B is
defined in the YAML file. If cluster 1 is missing a record from cluster
2, "B" will appear as the source cluster. However, on a re-run, cluster
1 knows nothing about namespace B, so exceptions will be thrown trying
to access it.

This commit solves by using the original namespace name and performing
an inverse mapping from B->A in this example. Note that this cannot be
guaranteed to be correct unless there is exactly one namespace being
used as the source, which is the most likely case.