Updated java client version to 7.2.1
Version changed to 7.2.1 solved a problem where Java serialized objects
were stored in the database and reading them back would cause a
ClassNotFoundException when the comparator attempted to deserialize
this. Version 7.2.1 does not return Java objects but instead BLOBs,
which the comparator can understand.

Also changed the command line parser to show an error and usage
information if an unrecognized options is passed.