Configurable timeouts
Reverted query / write / read policies to the default values as they
were in 1.2.2 but allowed them to be overridden in the config file.

The config file now has an optional "network" section which can have up
to 3 children: "read", "write" and "query". Each one of these can
specify any of the settings on a Policy class (socketTimeout,
totalTimeout, etc) which will override the defaults.

To show the network settings, use the "--verbose" (-V) flag which will
show these network settings on the console.

Rolled version to 1.2.5 and tagged.